Refactor front page feeds rendering
Web designers need to be able to modify the structure of the front page easily.
Having the HTML structure hard-coded deeply in PHP methods makes it difficult,
if not impossible. Therefore, this commit reimplements the way how data for
the front page feeds are generated.

Simply said, there must be no $OUTPUT used in local_moodleorg's locallib.php
code. It's the theme's duty to look after the actual representation. This
plugin must just provide all data required for rendering.