Include a fresh Apache conf for the moodle.org site
The new conf/moodle.org.conf contains bits from the current moodle.org live
site. It is supposed to be included into the vhost configuration for the
moodle.org site (or its development/test clones). Host specific configuration
(such as caching, https handling etc) is not included in this repository.