MDLSITE-2884 Fix the broken Apache conf file
The [NC,OR] added in the previous commit is not valid here. Also, as we do not
have /useful/index.php, we can't just redirect to /useful. So I'm fixing this
by explicitly allowing /useful/rss.php only and redirecting the rest to the
front page as it is now.