MDLSITE-4687 Fix the inability to update/revoke site registration
Email was not set for user accounts that represent registered sites and
are used as web service client account. This led to "user not fully set
up" error and it was not possible to update the registration or
unregister the site.

The patch set a dummy yet unique email address for these accounts.
Additionally, this starts to use the core API for creating the user
again rather than doing it manually.